+++ app/src/main/java/com/example/androidproject/rickshaw.java
@@ -28,8 +28,8 @@ public class rickshaw extends AppCompatActivity {
             //Toast.makeText(this,"Minimum Seat Requirment is 1 !!")
             E2.setError("Minimum Seat Requirment is 1 !!");
         } else {
-            result = Double.parseDouble((E1.getText().toString())) * 19.21;
-            Fresult = Double.parseDouble((E2.getText().toString())) * result;
+            result = Integer.parseInt((E1.getText().toString())) * 19.21;
+            Fresult = Integer.parseInt((E2.getText().toString())) * result;
 
             strin = "Dear User. Thankyou for using our app Fare Fair.\n\n The Total amount you have to pay to Auto driver is " + Fresult + "\n One have to pay the amount of " + result +
                     " \n\n  We have calculated the average Auto price of India. The calculation has it's basis on following \n\n Current Petrol Price = 99.12 rs/litre \n Average Auto Price = 19.21 Rs\n Total No of Peoples = " + E2.getText().toString();
